+++ src/main/java/com/kata/kata1/offers/IOffer.java
@@ -3,5 +3,5 @@ package com.kata.kata1.offers;
 import com.kata.kata1.Item;
 
 public interface IOffer {
-   Float calculatePrice(Item item,int itemQuantity);
+   double calculatePrice(Item item,int itemQuantity);
 }
