+++ src/main/java/com/kata/kata1/offers/PriceDropOffer.java
@@ -9,11 +9,19 @@ import lombok.RequiredArgsConstructor;
 public class PriceDropOffer implements IOffer {
 
     int offerUnit;
-    float offerPrice;
+    double offerPrice;
+
+    /**
+     * calculate price when there is special price rule applied for combination of items bought
+     *
+     * @param item whose price to be calculated
+     * @param itemQuantity quantity of same item scanned
+     * @return total price based on special price rule
+     */
 
     @Override
-    public Float calculatePrice(Item item,int itemQuantity) {
-        float specialPrice= (itemQuantity/offerUnit)*offerPrice + (itemQuantity%offerUnit)*item.getPrice();
+    public double calculatePrice(Item item,int itemQuantity) {
+        double specialPrice= (itemQuantity/offerUnit)*offerPrice + (itemQuantity%offerUnit)*item.getPrice();
         return specialPrice;
     }
 }
